+++ src/main/java/com/ssafy/hebees/dashboard/dto/response/TotalErrorsResponse.java
@@ -5,15 +5,12 @@ import java.time.LocalDateTime;
 
 @Schema(description = "총 에러 수 응답 DTO")
 public record TotalErrorsResponse(
-    Integer totalErrors, // 총 에러 수
+    Long totalErrors, // 총 에러 수
     LocalDateTime asOf // 기준 시각
 ) {
 
-    public static TotalErrorsResponse of(Integer totalErrors, LocalDateTime asOf) {
-        return new TotalErrorsResponse(
-            totalErrors,
-            asOf
-        );
+    public static TotalErrorsResponse of(Long totalErrors, LocalDateTime asOf) {
+        return new TotalErrorsResponse(totalErrors, asOf);
     }
 
 }
