+++ src/main/java/com/ssafy/hebees/dashboard/dto/response/TrendKeyword.java
@@ -2,7 +2,7 @@ package com.ssafy.hebees.dashboard.dto.response;
 
 public record TrendKeyword(
     String text, // 키워드 텍스트
-    Integer count, // 등장 횟수
+    Long count, // 등장 횟수
     Float weight // 가중치 ( = count - minCount) / max(1, maxCount - minCount)
 ) {
